Move v2.common and v2.inputset modules to top level
