+++ pyproject.toml
@@ -94,7 +94,7 @@ graphs = ["matplotlib"]
 "vasp.bands" = "aiida_vasp.workchains.v2.bands:VaspBandsWorkChain"
 "vasp.hybrid_bands" = "aiida_vasp.workchains.v2.bands:VaspHybridBandsWorkChain"
 "vasp.relax" = "aiida_vasp.workchains.v2.relax:VaspRelaxWorkChain"
-"vasp.neb" = "aiida_vasp.workchains.neb:VaspNEBWorkChain"
+"vasp.neb" = "aiida_vasp.workchains.v2.neb:VaspNEBWorkChain"
 "vasp.v2.vasp" = "aiida_vasp.workchains.v2.vasp:VaspWorkChain"
 "vasp.v2.converge" = "aiida_vasp.workchains.v2.converge:VaspConvergenceWorkChain"
 "vasp.v2.bands" = "aiida_vasp.workchains.v2.bands:VaspBandsWorkChain"
