+++ src/aiida_vasp/common/magmapping.py
