+++ src/aiida_vasp/workchains/__init__.py
@@ -1,23 +1,11 @@
-from .neb import VaspNEBWorkChain
-from .v2.bands import VaspBandsWorkChain, VaspHybridBandsWorkChain
-from .v2.common.builder_updater import *
-from .v2.converge import VaspConvergenceWorkChain
-from .v2.relax import VaspMultiStageRelaxWorkChain, VaspRelaxWorkChain
-from .v2.vasp import VaspWorkChain
+from .v2 import *
 
 __all__ = (
-    'VaspBandUpdater',
     'VaspBandsWorkChain',
-    'VaspBuilderUpdater',
-    'VaspConvUpdater',
     'VaspConvergenceWorkChain',
-    'VaspHybridBandUpdater',
     'VaspHybridBandsWorkChain',
     'VaspMultiStageRelaxWorkChain',
-    'VaspNEBUpdater',
     'VaspNEBWorkChain',
-    'VaspPresetConfig',
-    'VaspRelaxUpdater',
     'VaspRelaxWorkChain',
     'VaspWorkChain',
 )
